make image and file upload consistent with other components
