+++ src/components/ImageUpload/ImageUpload.js
@@ -1,0 +1,55 @@
+import { ErrorMessage, useFormik } from 'formik'
+import TextError from '../TextError/TextError'
+import { primaryStyles, textStyles } from '../../styles/styles'
+import { getMaxWidthClass } from '../../helpers/optionClasses'
+import imageUploadValidationSchema from "../../helpers/imageUploadValidationSchema"
+
+function ImageUpload(props) {
+  const { label, name, maxWidth, helper, required, ...rest } = props
+  const fieldMaxWidth = getMaxWidthClass(maxWidth);
+  const requiredClass = required ? primaryStyles.required : ''
+  const initialValues = {
+    file: null,
+  };
+  const onSubmit = (values) => {
+    // You can handle the file here, e.g., upload it to a server
+    console.log("Uploaded file:", values.file);
+  };
+  const validationSchema = imageUploadValidationSchema
+  const formik = useFormik({
+    initialValues,
+    validationSchema,
+    onSubmit,
+  });
+  return (
+    <div className={`${primaryStyles.wrapper} ${fieldMaxWidth} ${requiredClass} form-control`}>
+      <label
+        htmlFor={name}
+        className={textStyles.label}
+      >
+        {label} {required && <span className={textStyles.required}>*</span>}
+      </label>
+
+      {helper && <div className={textStyles.helper}>{helper}</div>}
+        <div className="mb-2">
+          <label htmlFor="file">Upload a Image:</label>
+          <input
+            type="file"
+            id="file"
+            name="file"
+            onChange={(event) => {
+              const selectedFile = event.currentTarget.files?.[0]; // Use optional chaining
+              formik.setFieldValue("file", selectedFile);
+            }}
+            onBlur={formik.handleBlur}
+          />
+          {formik.touched.file && formik.errors.file && (
+            <div className="text-red-500">{formik.errors.file}</div>
+          )}
+        </div>
+      <ErrorMessage name={name} component={TextError} />
+    </div>
+  )
+}
+
+export default ImageUpload
