--- src/helpers/useDebounceEffect.ts
@@ -1,13 +1,0 @@
-import { DependencyList, useEffect } from 'react';
-
-export function useDebounceEffect(fn: () => void, waitTime: number, deps: DependencyList) {
-  useEffect(() => {
-    const t = setTimeout(() => {
-      fn();
-    }, waitTime);
-
-    return () => {
-      clearTimeout(t);
-    };
-  }, deps);
-}
